+++ gambler/source/Fold/Pure.hs
@@ -17,7 +17,7 @@ module Fold.Pure
     {- *** Endpoints -} first, last,
     {- *** Extrema -} maximum, minimum, maximumBy, minimumBy,
 
-    {- * Conversion -} effectfulFold, nonemptyFold, shortcutNonemptyFold,
+    {- * Conversion -} effectfulFold, nonemptyFold, shortcutFold, shortcutNonemptyFold,
 
     {- * Utilities -} duplicate, premap, prefilter, predropWhile, drop, nest,
   )